app/requirements.txt
Flask
mysql-connector
openpyxl
waitress
